application-e/requirements.txt
Flask
Flask-Session
requests
PyJWT